requirements.txt
django~=2.2
pillow~=8.0.1
opencv-python~=4.4.0.46
django-ipware~=4.0.2
torch
torchvision
torchaudio
